+++ src/TraceabilitySystem.Infrastructure/Persistence/Repositories/ProcessRepository.cs
@@ -14,7 +14,7 @@ public class ProcessRepository : BaseRepository<Process>, IProcessRepository
 {
     public ProcessRepository(AppDbContext context) : base(context) { }
 
-    public override async Task<Process?> GetByIdAsync(long id, CancellationToken cancellationToken = default)
+    public override async Task<Process?> GetByIdAsync(int id, CancellationToken cancellationToken = default)
     {
         return await _dbSet
             .Include(p => p.ProcessParameters)
